Added fixes to azhpc-build for partially completed builds

+++ scripts/replace_nfs_with_anf.sh
@@ -19,6 +19,12 @@ chmod 777 $anf_mount_pt/apps
 chmod 777 $anf_mount_pt/data
 chmod 777 $anf_mount_pt/home
 #
-ln -s $anf_mount_pt/apps /apps
-ln -s $anf_mount_pt/data /data
-ln -s $anf_mount_pt/home /share/home
+if [ ! -L /apps ]; then
+    ln -s $anf_mount_pt/apps /apps
+fi
+if [ ! -L /data ]; then
+    ln -s $anf_mount_pt/data /data
+fi
+if [ ! -L /share/home ]; then
+    ln -s $anf_mount_pt/home /share/home
+fi
